tdf#130857 qt weld: Handle GtkViewPort child
The GtkViewPort doc [1] says:

> The GtkViewport widget acts as an adaptor class, implementing
> scrollability for child widgets that lack their own scrolling
> capabilities. Use GtkViewport to scroll child widgets such as GtkGrid,
> GtkBox, and so on.

For QScrollArea (which is created for "GtkScrolledWindow"
objects in .ui files), no special adaptor class is needed.

Therefore, just create a simple QWidget for "GtkViewport".

In theory, that widget could even be removed from the
hierarchy again later (e.g. in QtBuilder::tweakInsertedChild)
if there's a reason to.

[1] https://docs.gtk.org/gtk3/class.Viewport.html

Change-Id: I4d4f5164e73c8077aa4c5c65c080ee9d0e4676d0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178251
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins