tdf#130857 qt weld: Support dialog to choose UI variant
Declare support for the "View" -> "User Interface"
dialog that allows to choose a UI variant.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Currently, there's some black background behind
the images displaying the different UI variants,
which is presumably due to alpha not being handled
the way it should somewhere, This should be further
investigated at some point, but I don't consider
it blocking at the moment since the native Qt
dialogs are not used by default.
(Letting QtInstanceWidget::create_virtual_device
create a virtual device with a separate alpha
channel as SalInstanceWidget::create_virtual_device
does would result in the background being white
instead of black, as it is for the non-welded
dialog with qt6.)

Change-Id: I198750a27e13476b59569bda60949e286b567c5a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178254
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>