a11y: Use vcl::Window for VCLXAccessibleComponent validity check
Instead of checking for VCLXWindow reference being non-empty,
check whether the VclPtr<vcl::Window> ptr is non-empty.

The latter should never be true unless the former is
because VCLXAccessibleComponent::DisconnectEvents
gets called before both possible calls to

    m_xVCLXWindow.clear();

, and VCLXAccessibleComponent::DisconnectEvents
clears VCLXAccessibleComponent::m_xEventSource.

No change in behavior intended, but this further
reduces the use of the toolkit/UNO VCLXWindow
in favor of using the vcl::Window directly.

While at it, also drop the references to
tickets that don't provide any additional helpful
information:

"#122218#" is a reference to the StarDivision internal
StarOffice bugtracker, which is unavailable

"#i68079#" refers to [1], but also doesn't really
explain whether/why that check is needed:

From the initial description in that ticket:

> This is a follow-up of the issue 43665...  We still have the patch in
> ooo-build; it applies in IMPL_LINK( VCLXAccessibleComponent,
> WindowChildEventListener, VclSimpleEvent*, pEvent ) (few lines below the
> original problem).  Is it necessary (could it avoid a possible problem), or is
> it just bogus, and not needed?

Reply in comment 1:

> I'm not aware of any known issue, but I think this patch makes sense in
> order to avoid any potential problems. I will integrate this patch in
> one of the next childworkspaces.

[1] https://bz.apache.org/ooo/show_bug.cgi?id=68079

Change-Id: I14118cb95db4859ddac57d9646531cdbe0f967f2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178304
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins