tdf#150789 - FILEOPEN PPTX: fix text in SmartArt vertically off
Calculate correctly the textbox area of upArrowCallout and
downArrowCallout shapes in group shapes. In grouped list
text area does not cover the whole shape but just a part of it at the top.

In case of upArrowCallout and downArrowCallout the arrow size of the shape
is not included in the textbox area of the full shape.

Change-Id: If732305747c20da55bbd2896522c0b9c05cc4b4b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178343
Tested-by: Gabor Kelemen <gabor.kelemen.extern@allotropia.de>
Reviewed-by: Balazs Varga <balazs.varga.extern@allotropia.de>
Tested-by: Jenkins
(cherry picked from commit 9948c2ed03d227e178b9002dabc6edf34c32a5af)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178546
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>