tdf#164294 a11y: Don't rely on "creator" to determine child index
In OCommonAccessibleComponent::getAccessibleIndexInParent,
no longer rely on OCommonAccessibleComponent::m_aCreator
to have been set to the "creator" (i.e. the XAccessible whose
XAccessible::getAccessibleContext() created this object)
in order to determine the index of the object in its parent.

Instead, call XAccessible::getAccessibleContext on
the parent's child XAccessible objects and compare
that to `this`.

At least for SvxPixelCtlAccessible (and most likely other
child classes), OCommonAccessibleComponent::m_aCreator
isn't set (by a call to OCommonAccessibleComponent::lateInit),
so the logic would fail, and an incorrect index of -1
was returned.

Simplify the logic and no longer depend on
OCommonAccessibleComponent::m_aCreator in that
method. This also prepares for further decoupling
VCLXAccessibleComponent from the VCLXWindow toolkit/UNO class.
(The VCLXAccessibleComponent ctor currently calls
OCommonAccessibleComponent::lateInit with its VCLXWindow,
but overrides VCLXAccessibleComponent::getAccessibleIndexInParent
without making use of the "creator" there.)

Comments suggest that the previous logic was there to
avoid calling XAccessible::getAccessibleContext for
performance reasons.
If that's a concern for a particular subclass, overriding
OCommonAccessibleComponent::getAccessibleIndexInParent
in that subclass (and making use of whatever internal
details help to efficiently implement the method) seems
like a better and more reliable solution to me.

Change-Id: I29fb7cd42512a02fc1cc56835bb83f847e9ec0fd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178354
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins