a11y: Get text window doc accessible from Window on demand
Instead of storing a reference to the VCLXWindow (which is
a TextWindowPeer), which is the EditorWindow's accessible
as Document::m_xAccessible member, adjust Document::getAccessible
to get the accessible on demand via vcl::Window::GetAccessible
instead.

This also prepares for no more passing the VCLXWindow
to the ctor in an upcoming commit.

Change-Id: I498f90a456eda453508a4f1ed84983006594ea86
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178402
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins