a11y: No longer let accessibility module depend on toolkit
In the AccessibleBrowseBox ctor, no longer use the
VCLUnoHelper::GetInterface helper from the toolkit module,
but do the same inline.

(Assigning to AccessibleBrowseBoxBase::m_xFocusWindow
after the AccessibleBrowseBoxBase base class ctor has
been called looks a little odd to me as that one
would add a focus listener for that window if
already set/passed when called - but leave the
logic unchanged for now.)

With this, Library_acc no longer depends on the
toolkit module, so drop the dependency from the
make file.

Change-Id: I8e7759c2d92a344bef51d154df0045c78f0471fe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178435
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>