a11y: Make TabBar a11y independent of HAVE_FEATURE_SCRIPTING
Build the code for the TabBar a11y classes independent of
whether or not the scripting feature is enabled.

Otherwise, e.g. the tab bar showing sheets in Calc is inaccessible
in a `--disable-scripting` build, i.e. it e.g. doesn't show up
in Accerciser.

It was made conditional on HAVE_FEATURE_SCRIPTING in

    commit fb308c193bc9f0d07ad2cccd5a0b239f86cbe442
    Date:   Wed May 19 12:16:42 2021 +0200

        Fix --disable-scripting for DESKTOP build

, which doesn't give any explicit reason, so my
assumption is that it had some dependency relying
on that feature being enabled back then.

At least a local `--disable-scripting` build of mine
works fine with this enabled again now, so let's assume
whatever was the problem back then is solved by now.

Change-Id: I6ee2ea6823aaa30bf597678c2b3ee86b42111920
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178437
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>