tdf#130857 VclBuilder: Move more radio group logic to base class
Add a purely virtual WidgetBuilder::setRadioButtonGroup method
and call it for each of the entries in the map from
WidgetBuilder::processUIFile, in line with what is already done
for mnemonic widgets.
Move the existing logic for handling such a pair to the new
VclBuilder override, VclBuilder::setRadioButtonGroup.

An "actual" implementation for QtBuilder will be added in a
separate commit, just emit a warning there for now when the method
gets called.

Change-Id: I027278014bb18d9d5f3de8189cd2f515959bf3e7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178443
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins