tdf#130857 qt weld: Evaluate radio button groups
Add support for radio button groups by implementing
QtBuilder::setRadioButtonGroup.

Qt provides QButtonGroup [1] to explicitly group
buttons. Quoting from the QRadioButton doc [2]:

> Radio buttons are autoExclusive by default. If auto-exclusive is
> enabled, radio buttons that belong to the same parent widget behave as
> if they were part of the same exclusive button group. If you need
> multiple exclusive button groups for radio buttons that belong to the
> same parent widget, put them into a QButtonGroup.

To make sure exactly one button group exists and it gets
freed automatically, make the radio button whose ID is set for
the GtkRadioButton::group property [3] the parent (and thus the
owner) of the group.
Set a pointer of the group as a property for that button,
so it can be retrieved directly from the button without requiring
another map/... to keep track.

With this in place, the radio buttons in Writer's
"Tools" -> "Options" -> "LibreOfficeDev Writer"
-> "Mail Merge Email" -> "Server Authentication"
dialog are now grouped into 2 groups as expected in
a WIP branch, rather than all of them being in one group (because
they all have the same grid as a parent).
Support for that dialog will be declared in an
upcoming commit.

[1] https://doc.qt.io/qt-6/qbuttongroup.html
[2] https://doc.qt.io/qt-6/qradiobutton.html
[3] https://docs.gtk.org/gtk3/property.RadioButton.group.html

Change-Id: I60e8b5b525411c299ac50a8712361cc8572cc403
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178444
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>