tdf#164179 - Fix crash on BibEntry ComboBox select
One change in 7b99871635cd48c2a8a1d0afbd7afc60a45cc2ff (to
ToxAuthorityField) caused crashing bug in IdentifierHdl due to it
trying to call unsupported method on an input field to set its
value. This patch not only fixes the crash, but also moves the
volatile code into the SetFields method, which is also called in
the constructor of the Dialog, meaning if similar change were made
in the future, the crash would manifest immediately and hopefully
not pass to production.

Change-Id: I93abef0cdaa5c533523fbefd9514d1a287fd368b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178448
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179176
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>