tdf#130857 qt weld: Support "Alt Text" dialog
Declare support for the "Alt Text" dialog that
can e.g. be triggered as follows in Writer:

* "Insert" -> "OLE Object" -> "QR and Barcode"
* type anything, confirm
* select the QR code
* right click, select "Alt Text" in the context
  menu

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Currently, tab/focus behavior is a little odd,
as focus first moves to the scroll area that
the QScrollArea (created for the "GtkScrolledWindow"
object in the .ui file) before it moves inside the
QPlainTextEdit (created for the "GtkTextView" object
in the .ui file) contained therein.
That will be improved in an upcoming commit.

Change-Id: Ia94fdc8075a76af6a18b656a4c6b9168a408aee0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178483
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>