tdf#130857 tdf#162704 qt weld: Support QR/bar code dialog
Declare support for the "QR and Barcode" dialog
that can be triggered in Writer using "Insert" -> "OLE Object"
-> "QR/Barcode".

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

The native QPlainTextEdit also doesn't have the
problem of rendering artifacts as described in tdf#162704
for the vcl widget when used with qt6/kf6 and the Breeze
style.

Change-Id: I6263b8961d97141719e2b23dcc4467cda960d3e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178485
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>