tdf#130857 qrcodegen.ui: Define grid children in order
Switch the order in which the children of the grid in the
.ui file are defined so that the order matches the visual appearance,
which makes sure that tab focus order with the Qt-based VCL plugins is
correct as well in that dialog when using native Qt widgets, where
focus order is (by default) based on widget creation order.

See

    commit 02692566ad9fc7c3484f8581ffa0004cd4e43987
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 24 17:43:35 2024 +0200

        tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order

for more background.

Change-Id: I3995a50a34b2fc1515c5d64f5f1b6ff0dd83dbeb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178486
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>