CppunitTest_sw_apitests: remove leftover myNewGroup2.bau before the test
This fails for me in a tree that runs 'make check' and there was a
previous 'make check' previously.

Seems the trouble is that we have this C++ test creating an autotext
.bau file, but there is also a Java one, so if we're not lucky with the
order of these, the C++ one will fail with its hardcoded "must be 3 .bau
files in instdir" assertion.

Fix the immediate problem by removing the test .bau file before running
the C++ test.

A better, but more complex fix would be to create the test .bau file in
the user profile (and not in instdir) for both the Java and C++ cases,
and then none of these tests would influece other tests during 'make
check'.

Change-Id: I536d769db733515ec6a7d41fd7dbf601c4b582d4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178548
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178576
(cherry picked from commit 703cbc0bb75b0943cf5b95ff62c6fe4b4a028788)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178589
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>
Tested-by: Jenkins