a11y: Drop special menu/toolbar handling in AccessibleFactory
IIUC, this code path should never be reached:

The accessible for menu and menu floating windows
gets created in Menu::CreateAccessible. The
menu floating window reuses the Menu's accessible
and if the code path dropped in this commit was ever
reached for menu floating windows, it would be problematic,
see the commit messsage of

    commit b608604d0bbecdc24a17175374581b4e2d720658
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Dec 5 14:07:52 2024 +0000

        tdf#164093 tdf#157001 wina11y: Use vcl::Window's actual XAccessible

for more details.

Implementation for floating toolbar windows to have a null
XAccessible is implemented in ImplPopupFloatWin::CreateAccessible,
so that case also shouldn't get here (where a null XAccessibleContext
would have been returned).

Change-Id: Ie9a55b4120a6857e9cac534fa2bdfd3a0cf7d9c9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178584
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins