svtools a11y: Simplify Ruler::CreateAccessible
Ruler::mxAccContext is only used in this method
and nowhere else, so there's no need to have
this as a class member. (The vcl::Window
base class already holds a Reference to the
XAccessible, s. below.)

There's no need to explicitly call
vcl::Window::SetAccessible, because
Ruler::CreateAccessible gets called
by Window::GetAccessible, and that already
sets `mpWindowImpl->mxAccessible`, which
is what vcl::Window::SetAccessible also does.

Drop the

    // MT: Fixed compiler issue because the address from a temporary object was used.
    // BUT: Should it really be a Pointer, instead of const&???

comment for which it's unclear what it refers to now,
neither is it clear to me when looking at the
commit that originally introduced it,

    commit b755fb8c0f6b1282f62c12f378c0a5ecac64d490
    Author: Steve Yin <steve_y@apache.org>
    Date:   Mon Nov 25 16:15:58 2013 +0000

        Integrate branch of IAccessible2

Change-Id: Ic423d0fcc97894303c29097023742f2dabf00b79
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178585
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>