a11y: Move TabBar a11y classses to svtools, no longer use factory
Move AccessibleTabBar and related classes to
implement accessibility for the TabBar control
to svtools, where the TabBar code is also located.
(This matches how it's organized for the Ruler control
whose a11y class AccessibleRuler is also in svtools.)

This removes the last dependency of the accessibility
module on svtools, so drop that from accessibility/Library_acc.mk.

Instead of using the AccessibleFactory to
create an instance in TabBar::CreateAccessible,
just call the AccessibleTabBar ctor directly.

Change-Id: Id6a8852de930ffe7fe7509f84f33861d274120dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178588
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins