let insertModificationXcuFile accept "/" to indicate all nodes
insertModificationXcuFile takes a set of paths to include and a set of
paths to ignore from a given .xcu files

adapt this to allow an include path of "/" to match all contents of the
xcu.

insertModificationXcuFile is more useful than insertExtensionXcuFile for
the intended use case, because the latter inserts as an extension level
layer, and cannot override higher layers.

Change-Id: I8eabd9fb91184146a75b087cc31cee44f378d96d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178595
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>