tdf#164263 writerfilter: avoid exception when SAL_DISABLE_PRINTERLIST
This fixes a backported to 24.2.1 regression
from commit e6f75eb50bc5d1ebccda0c0f80a7b353d991106e

Apparently a blank value is allowed to be in RES_PAPER_BIN,
but it is not allowed to try to setPropertyValue as a blank string.

Lots of ways I could try to solve it.
-could just lump it in with staticDenylist,
 but that is comment-labeled as "read only values".
-could just wrap setPropertyValue in a try/catch,
 which is tempting because it is a generic fix
 for these otherwise silent failures,
 but that would hide potentially useful info
 behind a generic exception message.
-just do the one-off check like I first decided to do.

I don't see any unit test examples from the offending regression,
or anything else in SW. So I don't know how I can create
a unit test for this.

Change-Id: Ic2506dd4adb71f0e64d719363f6d261ffccf0151
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178620
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>