pdfium: remove duplication of converting a unicode string
This adds a universal getUnicodeString function that takes a
function call as a parameter, which is only a wrapped PDFium
function call that returns a unicode string. The rest of conversion
magic to OUString is then done by the function in one place.

This reduces a lot of duplication as we had to copy the way how
to convert the unicode string to OUString to every method.

Change-Id: Id951e2dd0c8b98837ca096cd5f800b8df0449fa0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178628
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178774
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>