[API CHANGE] a11y: Drop UNO service to create a11y contexts
Drop the com.sun.star.accessibility.GetStandardAccessibleFactoryService
UNO service that was used to create instances of
the a11y classes for (mainly vcl) widgets.

The service was needed to break a dependency cycle.
However, with

    Change-Id: Ib46c87446dc9121d3b8e735e0e5a40594da73cc5
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Dec 17 12:04:04 2024 +0100

        a11y: Merge accessibility module into vcl

and all the preparatory commits in place, there
is no more dependency cycle, as the a11y classes
for vcl widgets no longer depend on any modules
"above" vcl.

Therefore, drop the UNO service and the abstract
vcl::IAccessibleFactory class and other related
classes.

Make all methods in the AccessibleFactory class
(that was previously subclassing the abstract
IAccessibleFactory) static helper methods,
and call them directly.

The UNO service dropped in this commit was
introduced in

    commit 1af510e95147374e5289ae1c12f1f280931919f4
    Date:   Wed Feb 21 11:30:47 2024 +0200

        Create an UNO service to do the symbol lookup in toolkit::AccessibilityClient

and its documentation was already clearly stating
that it's only meant for LibreOffice internal use:

> /**
>   The toolkit module uses this to get a pointer to the AccessibleFactory from the acc module.
>   Because we have a dependency in our modules that goes the "wrong" way.
>
>   @since LibreOffice 24.8
>
>   @internal
>
>   ATTENTION: This is marked <em>internal</em> and does not
>   have the <em>published</em> flag, which means it is subject to
>   change without notice and should not be used outside the LibreOffice core.
> */

Change-Id: Ib97396a4bad486d9530361dd851ad3ee0f9681b8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178689
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins