tdf#164359 editeng RTF export: track unused parent styles recursively
Start edit mode for the shape in the document, double-click on a word
which has a "level 2" bullet associated with it -> crash.

This is a regression from c8b607b7c0096c58dc5187262bf0133dee728d50
(tdf#163883 editeng RTF export: fix broken offsets into the para style
table, 2024-11-19), the direct cause is that ImpEditEngine::WriteRTF()
maps all style pointers to a style ID and once we write the parent
property in the style table, we unconditionally dereference our search
result, because we know all styles should be in that map.

The root of the problem seems to be that once parent styles are marked
as used for the "only write used styles for the clipboard" case, those
parents themselves can have parents, and we didn't mark those as used.
Combined with the unconditional dereference, this leads to a crash. Fix
this by walking up the parent chain till we get no parent.

This assumes that importers create a valid document model, so there
can't be loops while walking the parent chain. If that would be a
problem in practice, we could check for that in the future.

Change-Id: I4b416e78f26bf455040d95ee8ed220cfe870e33a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178703
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit a7a81b6fbe37af938ce461e790fac517be032317)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178718
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>
(cherry picked from commit 5cba5e59da2102607048025a54a34aa6145dd30c)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178798
Reviewed-by: Aron Budea <aron.budea@collabora.com>