Related: tdf#164359 editeng RTF export: track unused follow styles recursively
Similar to commit a7a81b6fbe37af938ce461e790fac517be032317 (tdf#164359
editeng RTF export: track unused parent styles recursively, 2024-12-18),
the follows of a style has to be tracked recursively as well, to avoid a
crash.

The Impress UI doesn't seem to have a way to specify the parent/next
name of a style, but you can definitely create such a follow chain from
test code and probably this is also possible via macros.

Fix this similar to the parent case, except here handle when a style
sets itself as a follow: that's what the default Outline N styles do in
Impress.

Change-Id: If3847add02061fdb9ba1e3fbf7c1fc42e3866209
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178786
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>