tdf#164353 tdf#130857 qt weld: Keep extra msg dialog widget in place
Both, QMessageBox::setText, and
QMessageBox::setInformativeText call
QMessageBoxPrivate::setupLayout [1], and that one
deletes the QMessageBox's previous layout, creates
a new QGridLayout and sets that as the dialog's new
layout.

As a consequence, adding the widget that can hold
extra controls to the dialog's layout only once
by calling QtInstanceMessageDialog::addWidgetForExtraItems
from the QtInstanceMessageDialog ctor isn't sufficient,
but needs to be repeated each time the dialog layout
may have changed.
(Otherwise the widget is not contained in the layout,
and "randomly" placed in the dialog, so it can e.g.
overlap other UI elements in the dialog, see tdf#164353).

Therefore, rework QtInstanceMessageDialog::addWidgetForExtraItems
a bit,  rename it to QtInstanceMessageDialog::positionExtraControlsContainer
and call it also after updating texts.

[1] https://code.qt.io/cgit/qt/qtbase.git/tree/src/widgets/dialogs/qmessagebox.cpp#n290

Change-Id: Ib17b5867393a0694701c56b0ab74165021a8bc51
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178726
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>