tdf#161623 Fine tune menubar visibility when in LibreOffice full screen mode
Every once in a while, toggling LibreOffice full screen mode in a
native full screen window causes the menubar to be set to "hide"
instead of "autohide". So, expand the conditions for enabling
instead of disabling the menubar's visibility to include from
when a window is about to enter native full screen mode.

Also, just to be safe, always disable the menu items in the menubar
for LibreOffice full screen mode windows in case the menubar can't
be hidden while not in native full screen mode.

Change-Id: I4f1e8fc4a644cc13451690156b56662035321459
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178751
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>