cool#10630 lok doc sign: fix Impress sign line when creating directly
Once .uno:InsertSignatureLine gets dispatched, FuConstructRectangle gets
activated, the signature line (a graphic shape) gets inserted, but then
FuConstructRectangle is not deactivated in the "create directly" case,
so next clicks keep inserting new signature lines, while the expectation
is that "create directly" inserts one shape and selects it.

The "create directly" mode is the default for LOK, but you can trigger
it on the desktop with the KeyModifier=8192 (KEY_MOD1, i.e. Ctrl) UNO
command parameter, too. The normal mode would send the mouse button
down/up events to FuConstructRectangle and
sd::FuConstructRectangle::MouseButtonUp() would select the shape after a
drag to determine the position/size.

The "create directly" mode will insert instantly, so
sd::FuConstructRectangle::MouseButtonUp() is not called, so the shape is
not selected, so the shape insert mode is not deactivated. Fix this by
selecting the object for the create direct mode the same say as
FuConstructRectangle would do, at the place where e.g. text editing is
started for Impress textboxes.

Do this for rectangle shapes for now, but later other shapes might want
to do the same: e.g. curve insert doesn't have this problem, but it
seems line insert has the same problem -- that's not changed here.

Change-Id: I3ae85a07010f95b80a862e0cc631994113613eda
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178784
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>