tdf#157001 wina11y: Remove popup menu with its children
In AccTopWindowListener::windowClosed, drop the special
handling that was causing a popup menu (floating window)'s
accessible to not be removed from the objects that
the AccObjectWinManager manages.

It's unclear to me why that special handling was there
for accessibles with role
css::accessibility::AccessibleRole::POPUP_MENU, but might
have been some workaround for another issue fixed in the
meanwhile.

Removing the menu's children, but not the menu itself would
result in the menu's children not to be reinserted (and thus
no a11y listeners be registered) when the menu is shown again
after being closed, so that no focus events for the children
would be sent by the Windows a11y bridge.

Remove the menu's accessible as well and rely on the default
logic. This makes NVDA announce menu entries in Calc's
Organize Template menu also once the menu has been closed
and is then opened again.

Change-Id: Ie9cf96cad95215b9aeb0ebd7e4970da3c35b92c4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178825
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins