tdf#130857 qt weld: Create QSplitter for "GtkPaned"
GtkPaned [1] allows having two panes, allowing the
user to adjust the size using a handle.

In Qt, there's QSplitter [2] which would even support
having more than just two children and also handle it
in QtBuilder::replaceWidget.
This is e.g. used in the (not yet supported, but
seen in an experimental WIP branch) "File" -> "Open Remote"
dialog in Writer, where the GtkPaned has a GtkScrolledWindow
child that has a GtkTreeView child, and QTreeView
(created for the GtkTreeView) is a subclass of
QAbstractScrollArea, i.e. doesn't need a QScrollArea
parent in order to be scrollable.

Create a QSplitter for "GtkPaned" objects in .ui
files.

[1] https://docs.gtk.org/gtk3/class.Paned.html
[2] https://doc.qt.io/qt-6/qsplitter.html#replaceWidget

Change-Id: I7abaa1a0ca8e00046d5c3b1f2d4e00077b0aa360
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178850
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins