tdf#130857 qt weld: Support "Compress Image" dialog
Declare support for the "Compress Image" dialog that
can be triggered by selecting an image in Writer, right-clicking
and selecting the "Compress" context menu entry.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

This is the first supported dialog that has a slider ("GtkScale")
in it, for which support was added in previous commit

    Change-Id: I0d04b4db0b4531daba1e1110879acf85aff53aa3
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Dec 19 21:19:31 2024 +0100

        tdf#130857 qt weld: Add QtInstanceScale

        QSlider [1] is the Qt equivalent for GtkScale [2],
        so use it.

        [1] https://doc.qt.io/qt-6/qslider.html
        [2] https://docs.gtk.org/gtk4/class.Scale.html

Change-Id: I0cd7768f74d36272a9144efabb69864f167d3c23
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178875
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins