Revert "coverity#1430065 silence Uncaught exception"
This reverts commit 16d5cbfd38e57a6fccfda018ede2de7e4c448221.

Reason for revert:
The reverted commit was likely wrong, and was immediately obsoleted by
commit 7df7545fc9ac4aa55403998b2d65e1b147a9f037 (coverity#1430074
Uncaught exception, 2018-03-19), where the dtor of SwAutoFormatDlg got
a try...catch block.

The actual lines changed in the original commit were later modified in
commit c91faacfb7dd26fe7b7c93c7af5d4c330779a59f (make auto-format
dialog async, 2024-01-31), and this commit removes the use of
o3tl::default_delete, appeared in those subsequent changes.

Change-Id: Ibd634187d2930f3cc158aa0c36981ca8b327db91
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178876
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>