tdf#164396: Fix the minimum size for menu
It is not the full size of our minimal system requirements - it's
half of that: when the menu appears from the middle of the screen,
it only has half of the height.

Since the original commit 643a1492bd648fbd803ca86aca600cc2bdaf5819
(Add some fallback minimum size when screen size detection goes wrong,
2020-03-11) only mentioned misdetection of 0, this change should
be safe.

Change-Id: If62f4ee90dc062c37547fea4e72df2569d78f956
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178940
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>