tdf#130857 qt weld: Introduce QtInstanceMenuButton
Add new QtInstanceMenuButton class as the native Qt
weld::MenuButton implementation.

The actual logic of the class methods still needs to
be implemented, they all just trigger an assert when
called right now.

Create a QToolButton for "GtkMenuButton" objects in .ui files.

Let QtInstanceBuilder::weld_menu_button create an instance
of that new class and pass the button in the ctor.

Change-Id: Ie5ba0bad8e351554323e8f4ec1281b2c5ae8fbc3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178963
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins